refactor: move markdown impl into domain
